1) Spotbugs exclusions moved to config. We do not dirty the code with SuppressFBWarnings. 2) Added javadoc to HighAvailabilityPgConnectionImpl 3) Refactored construction methods in HighAvailabilityPgConnectionImpl. 4) Mocked tests moved to HighAvailabilityPgConnectionUnitTest with "fast" tag. 5) ClusterAwareTestBase renamed to PgConnectionAwareCluster 6) Implemented AutoCloseable for all Testcontainers wrappers. 6) Migrated tests to try-with-resources.
