+++ src/test/java/io/github/mfvanek/pg/connection/PrimaryHostDeterminerClusterTest.java
@@ -10,10 +10,8 @@
 
 package io.github.mfvanek.pg.connection;
 
-import io.github.mfvanek.pg.support.ClusterAwareTestBase;
+import io.github.mfvanek.pg.support.PgConnectionAwareCluster;
 import org.awaitility.Awaitility;
-import org.graalvm.compiler.core.common.SuppressFBWarnings;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.time.Duration;
@@ -27,42 +25,35 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @since 0.6.2
  */
 class PrimaryHostDeterminerClusterTest {
-    private final PrimaryHostDeterminer primaryHostDeterminer = new PrimaryHostDeterminerImpl();
-    private ClusterAwareTestBase postgresCluster;
 
-    @BeforeEach
-    protected void initCluster() {
-        this.postgresCluster = new ClusterAwareTestBase();
-    }
+    private final PrimaryHostDeterminer primaryHostDeterminer = new PrimaryHostDeterminerImpl();
 
     @Test
-    @SuppressFBWarnings(
-            value = "UWF_FIELD_NOT_INITIALIZED_IN_CONSTRUCTOR",
-            justification = "False positive"
-    )
     void correctPrimaryDetection() {
-        final PgConnection firstConnection = postgresCluster.getFirstPgConnection();
-        final PgConnection secondConnection = postgresCluster.getSecondPgConnection();
+        try (PgConnectionAwareCluster postgresCluster = new PgConnectionAwareCluster()) {
+            final PgConnection firstConnection = postgresCluster.getFirstPgConnection();
+            final PgConnection secondConnection = postgresCluster.getSecondPgConnection();
 
-        assertThat(primaryHostDeterminer.isPrimary(firstConnection))
-                .as("First connection is primary")
-                .isTrue();
+            assertThat(primaryHostDeterminer.isPrimary(firstConnection))
+                    .as("First connection is primary")
+                    .isTrue();
 
-        assertThat(primaryHostDeterminer.isPrimary(secondConnection))
-                .as("Second connection is not primary")
-                .isFalse();
+            assertThat(primaryHostDeterminer.isPrimary(secondConnection))
+                    .as("Second connection is not primary")
+                    .isFalse();
 
-        postgresCluster.stopFirstContainer();
+            postgresCluster.stopFirstContainer();
 
-        Awaitility
-                .await("Second node becomes primary")
-                .atMost(Duration.ofSeconds(120)) // on some systems promoting to primary could take up to minute and even more
-                .with()
-                .pollInterval(Duration.ofSeconds(2))
-                .until(() -> primaryHostDeterminer.isPrimary(secondConnection));
+            Awaitility
+                    .await("Second node becomes primary")
+                    .atMost(PgConnectionAwareCluster.MAX_WAIT_INTERVAL_SECONDS)
+                    .with()
+                    .pollInterval(Duration.ofSeconds(2))
+                    .until(() -> primaryHostDeterminer.isPrimary(secondConnection));
 
-        assertThat(primaryHostDeterminer.isPrimary(secondConnection))
-                .as("Second connection is primary")
-                .isTrue();
+            assertThat(primaryHostDeterminer.isPrimary(secondConnection))
+                    .as("Second connection is primary")
+                    .isTrue();
+        }
     }
 }
